setting/#1: SAA 적용 및 앱 진입 세팅
