+++ app/src/main/java/com/hsLink/hslink/HsuConnectApplication.kt
@@ -1,0 +1,25 @@
+package com.hsLink.hslink
+
+import android.app.Application
+import androidx.appcompat.app.AppCompatDelegate
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+
+@HiltAndroidApp
+class HsuConnectApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+
+        setTimber()
+        setDarkMode()
+    }
+
+    private fun setTimber() {
+        if (BuildConfig.DEBUG) Timber.plant(Timber.DebugTree())
+    }
+
+    private fun setDarkMode() {
+        AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
+    }
+}
