+++ app/src/main/java/com/hsLink/hslink/presentation/main/MainActivity.kt
@@ -1,0 +1,21 @@
+package com.hsLink.hslink.presentation.main
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import com.hsLink.hslink.core.designsystem.theme.HsLinkTheme
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            HsLinkTheme {
+                MainScreen()
+            }
+        }
+    }
+}
